Link self!
Now with the 'z80' target build and link sycek (ccheck, syc, z80test)
for Z80 and link binaries. Note that the binaries are not correct,
because they are larger than 64 kB. Added stubs for C library functions.
Fixed problem with variable being declared extern in the header file
and then declared non extern (and not initialized) not being defined
(stderr).

+++ lib/clib/include/string.h
@@ -28,7 +28,7 @@
 int asprintf(char **, const char *, ...);
 void *memcpy(void *, const void *, size_t) __attribute__((may_ignore_return));
 void *memmove(void *, const void *, size_t) __attribute__((may_ignore_return));
-void *memset(void *, int, unsigned) __attribute__((may_ignore_return));
+void *memset(void *, int, size_t) __attribute__((may_ignore_return));
 char *strchr(const char *, int);
 int strcmp(const char *, const char *);
 int strncmp(const char *, const char *, size_t);
